chore(ci): do not expect CSFLE/QE support in homebrew (#2420)
These aren't supported yet. Because my previous commit (1e6cef4a9170e)
pinned the version to an enterprise one, rather than a community server,
we were now trying to run the CSFLE smoke test but can't.

(We could also go back to testing with a community server, but in general
the idea of testing enterprise features in smoke tests is not a bad one.)