feat(shell-api): allow disabling "blocking cursor call" warnings MONGOSH-2674
Sometimes, users may decide to intentionally use the blocking behavior of these
methods. Since they cannot currently disable the warning, we give them a way to
do so by calling `cursor.disableBlockWarnings()`.

In order to make this a bit easier, we split the existing `AbstractCursor`
class into a `BaseCursor` class that contains logic that can be shared with
change stream cursors, and a `AbstractFiniteCursor` class that specifically
accounts for other existing cursor types.

As a side effect, this makes common cursor methods such as `.forEach()` and
`.map()` work for change stream cursors.