chore(shell-bson): split shell BSON logic into separate package (#2537)
This is admittedly a large refactor, but one that I would believe is generally
worth it:

- Moving the BSON print handling out of `service-provider-core` makes sense,
  as `service-provider-core` is generally intended to be a library
  providing an *interface*, not partial implementations.
- Moving the BSON classes out of `shell-api` is less architecturally important,
  but it does provide a little bit of additional encapsulation.
  It does, for example, force us to separate mongosh-specific `.help` property
  from core BSON constructor logic.
- A number of files in other packages were using `@mongosh/service-provider-core`
  only as a convenient way of getting the current `bson` package,
  which works but is really more of a hack and there's no good reason
  not to just use the `bson` package that the driver team ships
  directly.

This is also convenient on a broader basis, though, because it makes
mongosh's BSON constructors generally available for other projects with
similar needs to use. In Compass, we are avoiding this by using a Babel-based
"manual JavaScript evaluator", but not all projects have the same
technical requirements as Compass, and some may prefer the flexibility
of a full-fledged JS execution environment like mongosh does.