chore(ci): partially fix vscode test (#2538)
Right now, the VSCode test is failing in CI because the versions
of `@mongodb-js/devtools-connect` mismatch. We can align them the
same way we aligned the driver versions.

This does not turn the task green yet, but it is a necessary step
(the next one would be adjusting the VSCode source code for the
removal of the argument to `serviceProvider.close()` from c8e2697e).