chore(deps): update boxednode to 3.0.0, so we can support Node.js 24 MONGOSH-2995 (#2582)
* chore: Update boxednode to 3.0.0, so we can support Node.js 24

We are also upgrading the Node.js version we use for the driver
nightly tests to 24.11.1 (latest LTS version at the moment of
this commit)

* chore: Update our packages and the smoke tests to use also Node 24

* chore: Setup MSBuild for Windows builds (required for Node.js 24)

* chore: Upgrade node-gyp, when supported, so it works with clang-cl

Older versions of node-gyp use the /LCTG:INCREMENTAL flag, which
is not supported by clang-cl: it's only supported in the "standard"
msvc compiler.

* chore: Upgrade dependencies and use bundled npm on smoke tests

The latest Node.js version should provide the best npm version
with it. npm@10 is already older to what Node.js 24 provides
(npm@11).

* chore: Update to latest node-gyp. It should work with Node.js>=20

* chore: Upgrade all GHA to use the latest Node.js setup

* chore: Use bundled MSBuild in the GHA, it should still work

* chore: Rollback changes on .evergreen-nightly-driver