chore(ci): add all relevant Node.js 24 variants to evergreen for CI MONGOSH-2995 (#2586)
* chore: Update boxednode to 3.0.0, so we can support Node.js 24

We are also upgrading the Node.js version we use for the driver
nightly tests to 24.11.1 (latest LTS version at the moment of
this commit)

* chore: Update to latest node-gyp. It should work with Node.js>=20

* chore: Add a Node.js 24 variant to evergreen runs

We will be running all tests in both Node.js 20 and Node.js 24
with the current driver version. The nightly driver run that
happens every day should do the same with the nightly driver
without additional changes.

* chore: We should be using the npm version provided by Node.js

While older versions of npm had specific issues on Windows, the
latest npm@10 and npm@11 seem to be working fine, as they are
the ones we are using already in boxednode.

* chore: Use --jitless while we have the bugged 24.10 version of Node

* chore: In Node.js 22 and Node.js 24 __dirname does not exist

For some reason, Node.js 22 and Node.js 24 are more strict on
whether __dirname exists. Because import.meta.dirname always
exist, and must be equal, replace at the beginning of the script.

* chore: Update dependencies and update minimum Node.js version

Update to the minimum Node.js version that we have in the
devtools-toolchain (20.19.3)

* chore: Use process.cwd() instead of __dirname or import.meta

While this doesn't have the exact same semantic, it works
identical in all Node.js versions and we have control of
the cwd of the process, so this should be consistent and
work without ESM/CJS trics.

* chore: Path was somehow wrong, and it triggered using ESM

* chore: Overwriting __dirname just breaks Node 20.

* chore: use --no-experimental-strip-types in 24.11.1

* chore: Remove unused config (was not wired yet)

* chore: Rollback changes on .evergreen-nightly-driver

* chore: Fixes from PR comments

* chore: Remove the --jitless flag on v24

Evergreen is already working on providing an up-to-date
version of Node.js doesn't contain the JIT compiler
issue. We need JIT for WASM support.

* chore: Make Node.js 24 variants optional, so they don't block releases

Node.js 24 variants are going to be broken for a while, while we
work on the upgrade. It doesn't make much sense to keep them
running and blocking CI, so we will disable them by default and
only run them when a patch asks for it.

* chore: Update evergreen config so Node.js 24 is finally optional