fix(cli-repl): clear window title when mongosh exits MONGOSH-1233
Terminals do not provide an easy way to "push/pop" window titles,
so we're basically stuck with clearing the window title when mongosh
exits on a best-effort basis.

When somebody brings this up as an issue, we should still recommend
them to include something in their shell's .rc file to automatically
reset window titles after each shell command.