fix(ci): downgrade cron-tasks environment to Node.js 20/npm 10 (#2604)
chore: Downgrade cron-tasks environment to Node 20/npm 10

With npm 11, some dependencies related to the AWS driver
and Kerberos are not loaded properly. It also complains about
issues with prettier (where we use 2.x and 3.x).

While at some point we will need to update all these dependencies,
we will do it with the whole bump to Node.js 24, which requires other
work before that.