chore(e2e-tests): force wait for initialization before writing input (#2600)
In the past, we've had a number of flaky tests that followed a pattern
similar to this:

1. Start a test shell
2. Run a command using `executeLine()`
3. Make assertions about the command's output

The internal logic of `executeLine()` determines that it has seen
the end of the command's output by looking for the subsequent prompt
(similar to how a human being in front of the shell would determine
that the command has finished).

However, in the above scheme, there is a race condition that can
develop as follows:

1. The test shell process starts and prints part of its initial output
2. The `executeLine()` command remembers the current output position
   and sends the input to the process's stdin stream
3. The test shell prints its first prompt
4. The test reads the output from the shell and sees the prompt,
   assumes that the comamnd has finished
5. The test shell finishes evaluating the input and writes the
   corresponding output + a new prompt
6. The test ignores the remaining output and instead only treats
   the text up to the first prompt as the command output
7. An assertion made about the output fails, as the output is not
   coming from the command it was supposed to come from

In order to avoid the likelihood of introducing such race conditions,
this commit adds a guardrail to the input writing methods that
ensures that the shell has reached some form of known state that has
been awaited, such as seeing the first prompt or other specific output,
before it allows writing to the process's input.