chore(arg-parser): include zod to yargs conversion output in snapshot
Right now, we generate the yargs config based on the zod schema introduced
in 57b6adddbef88be65 as part of argument parsing.

Since in practice, the yargs config used by mongosh is static, we
can use currying to do that part of the call early and include
its result in the startup snapshot, removing runtime execution cost.